+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Changes
 
-* Added wallet generation from seed & import from seed on web sdk (#710)
+* Added wallet generation from seed & import from seed on web SDK (#710)
 * Add check for empty pay to id notes (#714).
 * [BREAKING] Refactored authentication out of the `Client` and added new separate authenticators (#718).
 * Re-exported RemoteTransactionProver in `rust-client` (#752).